# docs/swagger.yaml
basePath: /
definitions:
  dto.CheckPinBody:
    properties:
      pin:
        type: string
    required:
    - pin
    type: object
  dto.DashboardResponse:
    properties:
      balance:
        type: number
      total_expense:
        type: number
      total_income:
        type: number
    type: object
...
  dto.GetProfileResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      fullname:
        type: string
      id:
        type: integer
      phone_number:
        type: string
      photo_path:
        type: string
    type: object
  dto.GraphFilter:
    properties:
      end_date:
        description: '"2024-01-31"'
        type: string
      start_date:
        description: '"2024-01-01"'
        type: string
      type:
        description: '"income", "expense", "both"'
        type: string
    type: object
  dto.GraphPoint:
    properties:
      expense:
        description: ← total transfer
        type: number
      income:
        description: ← total topup
        type: number
      label:
        description: ← label sumbu X (tanggal/bulan)
        type: string
    type: object
  dto.GraphResponse:
    properties:
      filter:
        $ref: '#/definitions/dto.GraphFilter'
      points:
        items:
          $ref: '#/definitions/dto.GraphPoint'
        type: array
    type: object
...
  dto.PaginationMeta:
    properties:
      current_page:
        type: integer
      limit:
        type: integer
      total_data:
        type: integer
      total_pages:
        type: integer
    type: object
  dto.ReceiverListResponse:
    properties:
      meta:
        $ref: '#/definitions/dto.PaginationMeta'
      receivers:
        items:
          $ref: '#/definitions/dto.ReceiverResponse'
        type: array
    type: object
  dto.ReceiverResponse:
    properties:
      email:
        type: string
      fullname:
        type: string
      id:
        type: integer
      phone_number:
        type: string
      photo_path:
        type: string
      wallet_id:
        type: integer
    type: object
...
  dto.SwaggerDashboardResponse:
    properties:
      data:
        $ref: '#/definitions/dto.DashboardResponse'
      message:
        type: string
      success:
        type: boolean
    type: object
  dto.SwaggerGraphResponse:
    properties:
      data:
        $ref: '#/definitions/dto.GraphResponse'
      message:
        type: string
      success:
        type: boolean
    type: object
...
  dto.SwaggerProfileResponse:
    properties:
      data:
        $ref: '#/definitions/dto.GetProfileResponse'
      message:
        type: string
      success:
        type: boolean
    type: object
  dto.SwaggerReceiversResponse:
    properties:
      data:
        $ref: '#/definitions/dto.ReceiverListResponse'
      message:
        type: string
      success:
        type: boolean
    type: object
...
  dto.UpdatePasswordBody:
    properties:
      new_password:
        minLength: 6
        type: string
      old_password:
        minLength: 6
        type: string
    required:
    - new_password
    - old_password
    type: object
  dto.UpdatePinBody:
    properties:
      new_pin:
        type: string
      old_pin:
        type: string
    required:
    - new_pin
    type: object
...
  pkg.BaseResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
  pkg.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:9000
info:
  contact: {}
  description: Backend created by Koda using Gin
  license:
    name: MIT
  title: Backend E-WALLET
  version: "1.0"
paths:
...
  /users/dashboard:
    get:
      description: Mengambil informasi dashboard (balance, total income, total expense)
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil
          schema:
            $ref: '#/definitions/dto.SwaggerDashboardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get dashboard info
      tags:
      - Dashboard
  /users/dashboard/graph:
    get:
      description: Mengambil data graph transaksi dengan filter type dan date range
      parameters:
      - description: 'Filter: income / expense / both (default: both)'
        in: query
        name: type
        type: string
      - description: 'Tanggal mulai format: 2024-01-01 (default: 7 hari lalu)'
        in: query
        name: start_date
        type: string
      - description: 'Tanggal akhir format: 2024-01-31 (default: hari ini)'
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SwaggerGraphResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get dashboard graph
      tags:
      - Dashboard
  /users/password:
    patch:
      consumes:
      - application/json
      description: Memverifikasi Password user lama sebelum mengganti password baru
      parameters:
      - description: Update Password Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePasswordBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update password user
      tags:
      - Users
  /users/pin:
    patch:
      consumes:
      - application/json
      description: Mengubah PIN lama dengan varifikasi PIN yang baru
      parameters:
      - description: Update PIN Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePinBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update PIN user
      tags:
      - Users
  /users/pin/check:
    post:
      consumes:
      - application/json
      description: Memverifikasi PIN user sebelum melakukan transaksi
      parameters:
      - description: Check PIN Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CheckPinBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Check PIN user
      tags:
      - Users
  /users/profile:
    get:
      description: Mengambil data profile user yang sedang login
      produces:
      - application/json
      responses:
        "200":
          description: Profile berhasil diambil
          schema:
            $ref: '#/definitions/dto.SwaggerProfileResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get profile user
      tags:
      - Users
    patch:
      consumes:
      - multipart/form-data
      description: Update profile dengan atau tanpa upload foto
      parameters:
      - description: Nama lengkap
        in: formData
        name: fullname
        type: string
      - description: Nomor telepon
        in: formData
        name: phone_number
        type: string
      - description: Foto profile (opsional)
        in: formData
        name: photo
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update profile user
      tags:
      - Users
  /users/receiver:
    get:
      description: Mencari user lain sebagai penerima transfer dengan search dan pagination
      parameters:
      - description: Cari berdasarkan nama/email/phone
        in: query
        name: search
        type: string
      - description: 'Halaman (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Jumlah data per halaman (default: 7)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil
          schema:
            $ref: '#/definitions/dto.SwaggerReceiversResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Cari penerima transfer
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token dengan format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"